Remove tooling version and consolidate on the shared version.
